Move WorkspaceImporter out of workspace-utils
