+++ src/components/workspace-utils.js
@@ -9,7 +9,6 @@ import { icon, spinner } from 'src/components/icons'
 import { ValidatedInput } from 'src/components/input'
 import { MarkdownEditor, MarkdownViewer } from 'src/components/markdown'
 import Modal from 'src/components/Modal'
-import NewWorkspaceModal from 'src/components/NewWorkspaceModal'
 import TooltipTrigger from 'src/components/TooltipTrigger'
 import { Ajax, useReplaceableAjaxExperimental } from 'src/libs/ajax'
 import colors from 'src/libs/colors'
@@ -305,53 +304,6 @@ export const SnapshotInfo = ({
     ])
   ])
 }
-
-export const WorkspaceImporter = _.flow(
-  withDisplayName('WorkspaceImporter'),
-  withWorkspaces
-)(({ workspaces, refreshWorkspaces, onImport, authorizationDomain: ad, selectedWorkspaceId: initialWs, additionalErrors, ...props }) => {
-  const [selectedWorkspaceId, setSelectedWorkspaceId] = useState(initialWs)
-  const [creatingWorkspace, setCreatingWorkspace] = useState(false)
-
-  const selectedWorkspace = _.find({ workspace: { workspaceId: selectedWorkspaceId } }, workspaces)
-
-  return h(Fragment, [
-    h(WorkspaceSelector, {
-      workspaces: _.filter(ws => {
-        return Utils.canWrite(ws.accessLevel) &&
-          (!ad || _.some({ membersGroupName: ad }, ws.workspace.authorizationDomain))
-      }, workspaces),
-      value: selectedWorkspaceId,
-      onChange: setSelectedWorkspaceId,
-      ...props
-    }),
-    div({ style: { display: 'flex', alignItems: 'center', marginTop: '1rem' } }, [
-      h(ButtonPrimary, {
-        disabled: !selectedWorkspace || additionalErrors,
-        tooltip: Utils.cond([!selectedWorkspace, () => 'Select valid a workspace to import'],
-          [additionalErrors, () => Utils.summarizeErrors(additionalErrors)],
-          () => 'Import workflow to workspace'
-        ),
-        onClick: () => onImport(selectedWorkspace.workspace)
-      }, ['Import']),
-      div({ style: { marginLeft: '1rem', whiteSpace: 'pre' } }, ['Or ']),
-      h(Link, {
-        disabled: additionalErrors,
-        onClick: () => setCreatingWorkspace(true)
-      }, ['create a new workspace'])
-    ]),
-    creatingWorkspace && h(NewWorkspaceModal, {
-      requiredAuthDomain: ad,
-      onDismiss: () => setCreatingWorkspace(false),
-      onSuccess: w => {
-        setCreatingWorkspace(false)
-        setSelectedWorkspaceId(w.workspaceId)
-        refreshWorkspaces()
-        onImport(w)
-      }
-    })
-  ])
-})
 
 export const WorkspaceTagSelect = props => {
   const signal = useCancellation()
