+++ src/libs/workspace-utils.ts
@@ -19,6 +19,7 @@ export interface BaseWorkspaceInfo {
   name: string
   workspaceId: string
   cloudPlatform: string
+  authorizationDomain: string[]
 }
 
 export interface AzureWorkspaceInfo extends BaseWorkspaceInfo {
